sankalp-insight-node/requirements.txt
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
langdetect==1.0.9
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytest==8.3.3
Pillow==10.4.0